move Checklist editor files to @planx/components/Checklist/Editor
